# rest-app-openapi/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: Setlist Organizer API
  description: '''Setlist Organizer'' is a web application for organizing repertoires
    of musical bands.'
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0)
  version: 1.0.0
...
servers:
- url: http://localhost:8088
  description: Generated server url
...
paths:
  /repertoire:
    get:
      tags:
      - Track
      summary: Get information for all tracks based on their IDs
      operationId: tracks
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        "422":
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "200":
          description: A set of tracks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Track'
    put:
      tags:
      - Track
      summary: Update a track
      operationId: updateTrack
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Track'
        required: true
      responses:
        "404":
          description: Trying to update a non-existent track
          content: {}
        "400":
          description: Trying to update track with invalid fields
          content: {}
        "422":
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "200":
          description: Track(s) have been updated. Returns the number of tracks affected
          content:
            application/json:
              schema:
                type: integer
                format: int32
    post:
      tags:
      - Track
      summary: Create a new track
      operationId: createTrack
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Track'
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "400":
          description: An attempt to create track with invalid fields
          content: {}
        "422":
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "200":
          description: Track have been created. Returns the ID of the new track
          content:
            application/json:
              schema:
                type: integer
                format: int32
...
  /repertoire/import/excel:
    post:
      tags:
      - Track
      summary: Import information in the table 'Track' from Excel
      operationId: importTrackFromExcel
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
              - file
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        "422":
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "200":
          description: Track(s) have been imported. Returns the number of tracks imported.
          content:
            application/json:
              schema:
                type: integer
                format: int32
...
  /repertoire/{id}:
    get:
      tags:
      - Track
      summary: Get information for a single track identified by its unique ID
      operationId: getTrackById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "404":
          description: Trying to get a non-existent track
          content: {}
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        "422":
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "200":
          description: A track
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
    delete:
      tags:
      - Track
      summary: Delete a track
      operationId: deleteTrack
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "404":
          description: Trying to delete a non-existent track
          content: {}
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        "422":
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "200":
          description: Track(s) have been removed. Returns the number of tracks affected
          content:
            application/json:
              schema:
                type: integer
                format: int32
...
  /repertoire/fill:
    get:
      tags:
      - Track
      summary: Fill information for fake tracks based on their IDs
      operationId: tracksFake
      parameters:
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 1
      - name: language
        in: query
        required: false
        schema:
          type: string
          default: EN
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        "422":
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "200":
          description: A set of fake tracks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Track'
  /repertoire/export/xml:
    get:
      tags:
      - Track
      summary: Export information for all tracks based on their IDs to XML
      operationId: exportToXmlAlTracks
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        "422":
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "200":
          description: Successfully export to XML
          content:
            application/xml:
              schema:
                type: string
                format: binary
  /repertoire/export/excel:
    get:
      tags:
      - Track
      summary: Export information for all tracks based on their IDs to Excel
      operationId: exportToExcelAllTracks
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        "422":
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "200":
          description: Successfully export to Excel
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        details:
          type: array
          items:
            type: string
      description: ErrorResponse
    Track:
      required:
      - trackName
      type: object
      properties:
        trackId:
          type: integer
          description: ID of the track
          format: int32
          example: 1
        trackName:
          maxLength: 100
          minLength: 0
          type: string
          description: name of track
          example: New song
        trackBandId:
          type: integer
          description: ID of the band
          format: int32
          example: 1
        trackTempo:
          type: integer
          description: "track tempo, value in bpm"
          format: int32
          example: 120
        trackDuration:
          minimum: 0
          type: integer
          description: duration of the track play in ms
          format: int32
        trackDetails:
          maxLength: 2000
          minLength: 0
          type: string
          description: details about the track
        trackLink:
          maxLength: 255
          minLength: 0
          type: string
          description: the link to the track if it is posted on the web
          example: https://www.youtube.com/watch?v=Mp6W0IzLlW8&ab_channel=TheMuse
        trackReleaseDate:
          type: string
          description: track release date
          format: date
          example: 2020-12-01
      description: Track
...